+++ cakk-api/src/test/java/com/cakk/api/common/annotation/MockCustomUser.java
@@ -1,8 +1,9 @@
-package com.cakk.api.config;
+package com.cakk.api.common.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import org.springframework.security.test.context.support.WithSecurityContext;
+import com.cakk.api.common.config.WithMockCustomUserSecurityContextFactory;
 
 @Retention(RetentionPolicy.RUNTIME)
 @WithSecurityContext(factory = WithMockCustomUserSecurityContextFactory.class)
